:wrench: Mudança do peso da aresta de float para double.

+++ Grafo.c
@@ -4,7 +4,7 @@
 
 struct no {
     int rotulo;
-    float peso;
+    double peso;
     struct no *proximo;
 };
